backend/requirements.txt
# FastAPI and dependencies
fastapi
uvicorn
python-dotenv
fastapi>=0.68.0
uvicorn>=0.15.0
pygithub>=1.55
flake8>=4.0.1
radon>=5.1.0
pydantic>=1.9.0
requests>=2.27.1
python-dotenv>=0.19.2
# Optional AI integration
openai>=0.27.0